Add req 6

+++ desafio1.sql
@@ -7,7 +7,7 @@ USE SpotifyClone;
 CREATE TABLE planos(
     plano_id INT PRIMARY KEY AUTO_INCREMENT,
     plano VARCHAR(50) NOT NULL,
-    valor_plano DOUBLE NOT NULL
+    valor_plano DECIMAL(10, 2) NOT NULL
 ) engine = InnoDB;
 
 CREATE TABLE usuarios(
